[79-80] PR comments changes
