+++ stream-video-android/api/stream-video-android.api
@@ -2726,10 +2726,6 @@ public final class io/getstream/video/android/utils/ResultKt {
 	public static final synthetic fun onSuccessSuspend (Lio/getstream/video/android/utils/Result;Lkotlin/jvm/functions/Function2;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 }
 
-public final class io/getstream/video/android/utils/StringifyKt {
-	public static final fun stringify (Lio/getstream/video/android/utils/Result;Lkotlin/jvm/functions/Function1;)Ljava/lang/String;
-}
-
 public final class io/getstream/video/android/utils/Success : io/getstream/video/android/utils/Result {
 	public fun <init> (Ljava/lang/Object;)V
 	public final fun component1 ()Ljava/lang/Object;
